pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "retinalysis-enface"
dynamic = ["version"]
description = "Utilities for analyzing retinal enface images"
authors = [{ name = "Jose Vargas", email = "j.vargasquiros@erasmusmc.nl" }]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10, <3.12"
dependencies = [
    "numpy == 2.*",
    "pandas == 2.*",
    "scikit-learn == 1.*",
    "scipy == 1.*",
    "opencv-python == 4.*",
    "matplotlib == 3.*",
    "joblib == 1.*",
    "tqdm == 4.*",
    "Pillow == 9.*",
    "click == 8.*",
    "pydicom == 3.*",
]

[tool.setuptools]
packages = ["rtnls_enface"]

[project.scripts]
# Add any CLI scripts here if needed

[project.urls]
"Homepage" = "https://github.com/Eyened/retinalysis-enface"
"Bug Tracker" = "https://github.com/Eyened/retinalysis-enface/issues"

[tool.setuptools_scm]
# can be empty if no extra settings are needed, presence enables setuptools-scm 